+++ 박수정/interest/src/scripts/initializationInput.js
@@ -1,0 +1,4 @@
+export function initializationInput(el) {
+    el.value = '';
+    el.focus();
+}
