+++ 박수정/interest/src/scripts/setElement.js
@@ -1,0 +1,3 @@
+export const container = document.querySelector('.input__container');
+export const listWrap = container.querySelector('ul');
+export const input = container.querySelector('input');
